refactor(aws): move CDK under server/aws and split handlers/context
Relocate aws/resources to server/aws/resources and adopt handlers/ +
context/ layout for Lambda entries, updating CDK paths and READMEs.
