--- aws/resources/lib/util/paths.ts
@@ -1,10 +1,0 @@
-import * as path from 'path';
-
-/** Absolute path to aws/resources (CDK is run from this directory). */
-export function getResourcesRoot(): string {
-  return process.cwd();
-}
-
-export function handlerEntry(handlerFile: string): string {
-  return path.join(getResourcesRoot(), 'lib/handlers', handlerFile);
-}
