+++ server/aws/handlers/authz.ts
@@ -1,18 +1,18 @@
-import { createIssueRouter } from '@trustknots/server-core/routes/issue'
+import { createAuthzRouter } from '@trustknots/server-core/routes/authz'
 import { Hono } from 'hono'
 import { HTTPException } from 'hono/http-exception'
 import { handle } from 'hono/aws-lambda'
-import { createVcknotsContext, getBaseUrl } from '../util/vcknots-context'
+import { createVcknotsContext, getBaseUrl } from '../context/vcknots-context.js'
 
 const options = {
-  // TODO: AWS 用の provider（DynamoDB/KMS など）が揃ったらここに差し替える
+  // TODO: @trustknots/aws provider（DynamoDB/KMS など）が揃ったらここに差し替える
 }
 
 const context = createVcknotsContext(options)
 const baseUrl = getBaseUrl()
 
 const app = new Hono()
-app.route('/', createIssueRouter(context, baseUrl))
+app.route('/', createAuthzRouter(context, baseUrl))
 app.notFound((c) => c.json({ error: 'Not Found' }, 404))
 app.onError((err, c) => {
   if (err instanceof HTTPException) return err.getResponse()
