+++ server/aws/resources/.gitignore
