+++ server/aws/resources/.npmignore
