+++ server/aws/resources/bin/resources.ts
