+++ server/aws/resources/lib/construct/data/data-stores.ts
