+++ server/aws/resources/lib/resources-stack.ts
