+++ server/aws/resources/scripts/.env.example
