+++ server/aws/resources/tsconfig.json
