+++ server/aws/tsconfig.json
@@ -1,0 +1,7 @@
+{
+	"extends": "../../tsconfig.json",
+	"compilerOptions": {
+		"outDir": "./lib"
+	},
+	"include": ["./handlers", "./context"]
+}
